Bug #1757282 soc: sprd: domain: fix pclint warning
[root cause  ]para type
[changes     ]para type
[side effects]yes
[self test   ]:no
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:no
[is there dependence]:no
[confirm dependent commit]:no
[board]:l3 l5 l6
[test case]:power on
[reviewers   ]eddy.wei yangyang.liu joey.wu lu.liu1
[change_type ] feature_bugfix
[tag_product ] common

fix pclint warning

+++ drivers/soc/sprd/domain/sprd_camsys_domain.h
@@ -36,7 +36,7 @@ struct register_gpr {
 struct camsys_power_info;
 
 struct camsys_power_ops {
-	int (*sprd_campw_init)(struct platform_device *pdev, struct camsys_power_info *pw_info);
+	long (*sprd_campw_init)(struct platform_device *pdev, struct camsys_power_info *pw_info);
 	int (*sprd_cam_domain_eb)(struct camsys_power_info *domain);
 	int (*sprd_cam_domain_disable)(struct camsys_power_info *domain);
 	int (*sprd_cam_pw_on)(struct camsys_power_info *domain);
